Move common test helper code out of individual test files
