+++ src/tests/unit/common/visualization-toggle-props-builder.ts
@@ -1,0 +1,12 @@
+import { VisualizationToggleProps } from '../../../common/components/visualization-toggle';
+import { BaseDataBuilder } from './base-data-builder';
+export class VisualizationTogglePropsBuilder extends BaseDataBuilder<VisualizationToggleProps> {
+    constructor() {
+        super();
+        this.data = {
+            checked: false,
+            disabled: false,
+            visualizationName: null,
+        } as VisualizationToggleProps;
+    }
+}
